Increase heartbeat timeout from 15 seconds to 1 minutes (#29)
* Increase heartbeat timeout from 15 seconds to 5 minutes

We're processing over 400 queries per second at peak times now. About half of which
are updates. Increasing the heartbeat interval should cut that dramatically

Comments from the PR review:

I looked around to make myself comfortable with this.

First thing was to check where HEARTBEAT_INTERVAL is used, looks like 4 files:

WorkflowDbPersistenceFactoryIT: test file, not going to look further.

InitializedDbPersistence: This file implements and starts the heartbeat thread that is run by workflow. My only concern here is that in the thread, we first sleep for HEARTBEAT_INTERVAL, then actually heartbeat. So, if there is a short job, is it possible that it will finish before the interval is up, never heartbeat, and the column in the DB will be null? Will that cause problems somewhere? To answer my own question, the column has a null: false constraint and when we create the attempt we always set the heartbeat column.

DbPersistence: Where the constant is defined, no other usages.

WorkflowQueries: This file contains the interesting portion of the usages. Going into this we were fairly certain that the heartbeat column was only used to represent a "DIED UNCLEAN" or "Timed Out" status. The goal being that if your job was stuck in a wonky state, it would become clear as it missed heartbeats and you could cancel it. The getProcessStatus function uses HEARTBEAT_INTERVAL for this exact purpose and is called in a couple places:

DiedUnclean: This monitor checks the status of workflows and alerts if something is in a TIMED_OUT status. I'm not particularly concerned about this because I don't believe we run the workflow_monitor in our deployment. However, if people are running workflow_monitor, I'd like to note that the default missed heartbeat limit is 20. Increasing the HEARTBEAT_INTERVAL to 5 minutes would mean that a workflow attempt would have to miss heartbeats for 100 minutes before this alert fires. Since it appears we don't use the workflow_monitor and the number of missed heartbeats for this alert is configurable with the missed_heartbeats_limit param, I'm fine letting this be.
getProcessStatus is referenced in a few other files: Assertions, WorkflowJSON, ExecutionController, ApplicationController. I'm not going to follow these usages, but from a quick pass they all look like they are either getting the status to display in the UI or confirming if the attempt is alive or dead before allowing an action. Increasing the HEARTBEAT_INTERVAL should just be increasing the number of minutes before an attempt is declared dead.
All that being said, I would prefer to change the HEARTBEAT_INTERVAL to 1 minute as opposed to 5 minutes. It means that attempts would generally be TIMED_OUT in 4 minutes instead of 20 minutes (4x multiplier: constant, usage) and it will still cut down on the number of updates.