Stop displaying counters in the UI
Users can still click through the job history server if they want look
at the counters for a specific job. If people complain loudly, then maybe
we could query the job history server for the counters whenever someone
vists a page.

This is part of a longer-term plan to stop persisting counters in workflow
DB so we have fewer inserts overall. We think this will make it possible
for us to migrate workflow DB off of custom MySQL setup we have and onto
CloudSQL.